+++ src/main/kotlin/com/github/marcospereira/day6/Day6.kt
@@ -1,5 +1,6 @@
-package com.github.marcospereira
+package com.github.marcospereira.day6
 
+import com.github.marcospereira.Day
 import java.util.*
 
 /**
