+++ src/main/kotlin/com/github/marcospereira/day8/Day8.kt
@@ -1,0 +1,16 @@
+package com.github.marcospereira.day8
+
+import com.github.marcospereira.Day
+
+/**
+ * Created by francois on 2016-02-07.
+ */
+class Day8 : Day() {
+    override fun part1(): Any? {
+        throw UnsupportedOperationException()
+    }
+
+    override fun part2(): Any? {
+        throw UnsupportedOperationException()
+    }
+}
